create a set of goals targeting unification logic directly
Especially around lifetimes
