+++ src/test.rs
@@ -116,6 +116,7 @@ mod cycle;
 mod implied_bounds;
 mod impls;
 mod projection;
+mod unify;
 
 #[test]
 fn struct_wf() {
@@ -180,258 +181,6 @@ fn generic_trait() {
             Int: Eq<Uint>
         } yields {
             "No possible solution"
-        }
-    }
-}
-
-#[test]
-fn normalize_under_binder() {
-    test! {
-        program {
-            struct Ref<'a, T> { }
-            struct I32 { }
-
-            trait Deref<'a> {
-                type Item;
-            }
-
-            trait Id<'a> {
-                type Item;
-            }
-
-            impl<'a, T> Deref<'a> for Ref<'a, T> {
-                type Item = T;
-            }
-
-            impl<'a, T> Id<'a> for Ref<'a, T> {
-                type Item = Ref<'a, T>;
-            }
-        }
-
-        goal {
-            exists<U> {
-                forall<'a> {
-                    Ref<'a, I32>: Deref<'a, Item = U>
-                }
-            }
-        } yields {
-            "Ambiguous"
-        }
-
-        goal {
-            exists<U> {
-                forall<'a> {
-                    Normalize(<Ref<'a, I32> as Deref<'a>>::Item -> U)
-                }
-            }
-        } yields {
-            "Unique; substitution [?0 := I32], lifetime constraints []"
-        }
-
-        goal {
-            forall<'a> {
-                exists<U> {
-                    Ref<'a, I32>: Id<'a, Item = U>
-                }
-            }
-        } yields {
-            "Ambiguous"
-        }
-
-        goal {
-            forall<'a> {
-                exists<U> {
-                    Normalize(<Ref<'a, I32> as Id<'a>>::Item -> U)
-                }
-            }
-        } yields {
-            "Unique; substitution [?0 := Ref<'!1_0, I32>], lifetime constraints []"
-        }
-
-        goal {
-            exists<U> {
-                forall<'a> {
-                    Normalize(<Ref<'a, I32> as Id<'a>>::Item -> U)
-                }
-            }
-        } yields {
-            "Unique; for<?U0> { \
-             substitution [?0 := Ref<'^0, I32>], \
-             lifetime constraints [InEnvironment { environment: Env([]), goal: '^0 == '!1_0 }] \
-             }"
-        }
-    }
-}
-
-#[test]
-fn unify_quantified_lifetimes() {
-    test! {
-        program {
-        }
-
-        // Check that `'a` (here, `'^0`) is not unified
-        // with `'!1_0`, because they belong to incompatible
-        // universes.
-        goal {
-            exists<'a> {
-                forall<'b> {
-                    'a = 'b
-                }
-            }
-        } yields {
-            "Unique; for<?U0> { \
-             substitution [?0 := '^0], \
-             lifetime constraints [InEnvironment { environment: Env([]), goal: '^0 == '!1_0 }] \
-             }"
-        }
-
-        // Similar to the previous test, but indirect.
-        goal {
-            exists<'a> {
-                forall<'b> {
-                    exists<'c> {
-                        'a = 'c,
-                        'c = 'b
-                    }
-                }
-            }
-        } yields {
-            "Unique; for<?U0> { \
-             substitution [?0 := '^0, ?1 := '!1_0], \
-             lifetime constraints [InEnvironment { environment: Env([]), goal: '^0 == '!1_0 }] \
-             }"
-        }
-    }
-}
-
-#[test]
-fn equality_binder() {
-    test! {
-        program {
-            struct Ref<'a, T> { }
-        }
-
-        // Check that `'a` (here, `'?0`) is not unified
-        // with `'!1_0`, because they belong to incompatible
-        // universes.
-        goal {
-            forall<T> {
-                exists<'a> {
-                    for<'c> Ref<'c, T> = Ref<'a, T>
-                }
-            }
-        } yields {
-            "Unique; for<?U1> { \
-                 substitution [?0 := '^0], \
-                 lifetime constraints [InEnvironment { environment: Env([]), goal: '!2_0 == '^0 }] \
-             }"
-        }
-    }
-}
-
-#[test]
-fn mixed_indices_unify() {
-    test! {
-        program {
-            struct Ref<'a, T> { }
-        }
-
-        goal {
-            exists<T> {
-                exists<'a> {
-                    exists<U> {
-                        Ref<'a, T> = Ref<'a, U>
-                    }
-                }
-            }
-        } yields {
-            "Unique; for<?U0,?U0> { \
-                 substitution [?0 := '^0, ?1 := ^1, ?2 := ^1], \
-                 lifetime constraints []\
-             }"
-        }
-    }
-}
-
-#[test]
-fn mixed_indices_match_program() {
-    test! {
-        program {
-            struct S { }
-            struct Bar<'a, T, U> { }
-            trait Foo {}
-            impl<'a> Foo for Bar<'a, S, S> {}
-        }
-
-        goal {
-            exists<T> {
-                exists<'a> {
-                    exists<U> {
-                        Bar<'a, T, U>: Foo
-                    }
-                }
-            }
-        } yields {
-            "Unique; for<?U0> { \
-                 substitution [?0 := '^0, ?1 := S, ?2 := S], \
-                 lifetime constraints [] \
-             }"
-        }
-    }
-}
-
-#[test]
-fn mixed_indices_normalize_application() {
-    test! {
-        program {
-            struct Ref<'a, T> { }
-            trait Foo {
-                type T;
-            }
-
-            impl<U, 'a> Foo for Ref<'a, U> {
-                type T = U;
-            }
-        }
-
-        goal {
-            exists<T> {
-                exists<'a> {
-                    exists<U> {
-                        Normalize(<Ref<'a, T> as Foo>::T -> U)
-                    }
-                }
-            }
-        } yields {
-            "Unique; for<?U0,?U0> { substitution [?0 := '^0, ?1 := ^1, ?2 := ^1], "
-        }
-    }
-}
-
-#[test]
-fn mixed_indices_normalize_gat_application() {
-    test! {
-        program {
-            struct Either<T, U> { }
-            struct Ref<'a, T> { }
-            trait Foo {
-                type T<X>;
-            }
-
-            impl<U, 'a> Foo for Ref<'a, U> {
-                type T<X> = Either<X, U>;
-            }
-        }
-
-        goal {
-            exists<T, 'a, U, Y, X> {
-                Normalize(<Ref<'a, T> as Foo>::T<X> -> Either<U, Y>)
-            }
-        } yields {
-            // Our GAT parameter <X> is mapped to ?0; all others appear left to right
-            // in our Normalize(...) goal.
-            "Unique; for<?U0,?U0,?U0> { \
-                substitution [?0 := ^0, ?1 := '^1, ?2 := ^2, ?3 := ^0, ?4 := ^2], "
         }
     }
 }
