+++ src/test/impls.rs
@@ -245,77 +245,3 @@ fn normalize_rev_infer_gat() {
         }
     }
 }
-
-/// Basic tests of region equality: we generate constraints.
-#[test]
-fn region_equality() {
-    test! {
-        program {
-            trait Eq<T> { }
-            impl<T> Eq<T> for T { }
-
-            struct Unit { }
-            struct Ref<'a, T> { }
-        }
-
-        goal {
-            forall<'a, 'b> {
-                Ref<'a, Unit>: Eq<Ref<'b, Unit>>
-            }
-        } yields {
-            "Unique; substitution [],
-                     lifetime constraints \
-                     [InEnvironment { environment: Env([]), goal: '!1_1 == '!1_0 }]
-                     "
-        }
-
-        goal {
-            forall<'a> {
-                exists<'b> {
-                    Ref<'a, Unit>: Eq<Ref<'b, Unit>>
-                }
-            }
-        } yields {
-            "Unique; substitution [?0 := '!1_0], lifetime constraints []"
-        }
-    }
-}
-
-/// Tests of region equality and "foralls" -- we generate contraints that are sometimes
-/// not solvable.
-#[test]
-fn forall_equality() {
-    test! {
-        program {
-            trait Eq<T> { }
-            impl<T> Eq<T> for T { }
-
-            struct Unit { }
-            struct Ref<'a, T> { }
-        }
-
-        goal {
-            // A valid equality; we get back a series of solvable
-            // region constraints, since each region variable must
-            // refer to exactly one placeholder region, and they are
-            // all in a valid universe to do so (universe 4).
-            for<'a, 'b> Ref<'a, Ref<'b, Unit>>: Eq<for<'c, 'd> Ref<'c, Ref<'d, Unit>>>
-        } yields {
-            "Unique; substitution [], lifetime constraints []"
-        }
-
-        goal {
-            // Note: this equality is false, but we get back successful;
-            // this is because the region constraints are unsolvable.
-            //
-            // Note that `?0` (in universe 2) must be equal to both
-            // `!1_0` and `!1_1`, which of course it cannot be.
-            for<'a, 'b> Ref<'a, Ref<'b, Ref<'a, Unit>>>: Eq<
-                for<'c, 'd> Ref<'c, Ref<'d, Ref<'d, Unit>>>>
-        } yields {
-            "Unique; substitution [], lifetime constraints [
-                 InEnvironment { environment: Env([]), goal: '!1_1 == '!1_0 }
-             ]"
-        }
-    }
-}
